Fix : corrections bugs sur taille des notices

+++ core/src/main/java/fr/abes/periscope/service/NoticeStoreService.java
@@ -5,6 +5,7 @@ import fr.abes.periscope.repository.NoticeRepository;
 import fr.abes.periscope.util.NoticeMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessResourceFailureException;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -42,11 +43,19 @@ public class NoticeStoreService {
     }
 
     public Iterable<NoticeSolrExtended> saveList(List<NoticeSolrExtended> notice) {
-        return noticeRepository.saveAll(notice);
+        try {
+            if (notice.size() > 0)
+                return noticeRepository.saveAll(notice);
+        } catch (DataAccessResourceFailureException ex) {
+            log.error("Erreur d'indexation notice : " + ex.getMessage());
+        }
+            return null;
+
     }
 
     public void deleteList(List<NoticeSolrExtended> notice) {
-        noticeRepository.deleteAll(notice);
+        if (notice.size() > 0)
+            noticeRepository.deleteAll(notice);
     }
 
     public void delete(NoticeSolrExtended notice) {
